Move /auth tests to /auth/api_test.go
