+++ bcda/main.go
@@ -110,14 +110,14 @@ func setUpApp() *cli.App {
 				go func() { log.Fatal(srv.ListenAndServe()) }()
 
 				auth := &http.Server{
-					Handler:      NewAPIRouter(),
+					Handler:      NewAuthRouter(),
 					ReadTimeout:  time.Duration(utils.GetEnvInt("API_READ_TIMEOUT", 10)) * time.Second,
 					WriteTimeout: time.Duration(utils.GetEnvInt("API_WRITE_TIMEOUT", 20)) * time.Second,
 					IdleTimeout:  time.Duration(utils.GetEnvInt("API_IDLE_TIMEOUT", 120)) * time.Second,
 				}
 
 				api := &http.Server{
-					Handler:      NewAuthRouter(),
+					Handler:      NewAPIRouter(),
 					ReadTimeout:  time.Duration(utils.GetEnvInt("API_READ_TIMEOUT", 10)) * time.Second,
 					WriteTimeout: time.Duration(utils.GetEnvInt("API_WRITE_TIMEOUT", 20)) * time.Second,
 					IdleTimeout:  time.Duration(utils.GetEnvInt("API_IDLE_TIMEOUT", 120)) * time.Second,
